+++ Gnuciverba/Gnuciverba.pro
@@ -1,0 +1,8 @@
+TEMPLATE = app
+CONFIG += console
+CONFIG -= app_bundle
+CONFIG -= qt
+
+SOURCES += \
+        main.c \
+    Libreria.c
